More parsing stuff completed
todo
Options section indicated in HW file

Compute and display the size of the payload carried in the segment.

Use getservbyport to translate numbers found in the SOURCE PORT and
DESTINATION PORT fields to service names.

combining and function calls i should have bounds checks to make sure
what is being passed into the function is of valid size